projects/SimEngine/requirements.txt
nuplan-devkit @ git+https://github.com/motional/nuplan-devkit/@nuplan-devkit-v1.2
aioboto3==10.4.0
boto3==1.24.59
cachetools==5.2.0
fiona==1.8.21
geopandas>=0.12.1
hydra-core==1.2.0
matplotlib>=3.9.4
ninja>=1.13.0
numpy==1.23.0
opencv-python==4.10.0.84
pandas>=2.3.3
pillow==10.4.0
pip>=26.0
pyarrow>=21.0.0
pyogrio==0.4.1
pyquaternion>=0.9.5
pytest==7.1.3
rasterio==1.3.2
ray==2.48.0
retry==0.9.2
scipy>=1.13.1
setuptools==65.5.1
shapely>=2.0.0
sqlalchemy==1.4.27
tqdm>=4.67.2
urllib3==1.26.12
wheel>=0.46.3
psutil==6.1.0
imageio==2.36.1
aiofiles==24.1.0
pytorch-lightning==2.2.1
pyinstrument==5.0.0